chore: move agent related fixtures and linting
