+++ beamer/tests/conftest.py
@@ -1,21 +1,9 @@
 # flake8: noqa: E402
-import stat
 from dataclasses import dataclass
-from typing import cast
 
 import ape
-import eth_account
 import pytest
-
-import beamer.agent.chain
-import beamer.agent.metrics
-from beamer.agent.agent import Agent
-from beamer.agent.config import ChainConfig, Config
-from beamer.agent.relayer import get_relayer_executable
-from beamer.agent.util import TokenChecker
-from beamer.contracts import ContractInfo, DeploymentInfo
 from beamer.tests.util import alloc_accounts
-from beamer.typing import URL, BlockNumber, ChainId, TransferDirection
 
 
 @dataclass(frozen=True)
@@ -30,7 +18,9 @@ class Contracts:
 # ape local account, to be used for fulfilling requests.
 # The private key here corresponds to the 10th account ganache creates on
 # startup.
-_LOCAL_ACCOUNT = ape.accounts.test_accounts[-1]
+@pytest.fixture
+def local_account():
+    return ape.accounts.test_accounts[-1]
 
 
 @pytest.fixture
@@ -145,7 +135,14 @@ def forward_state():
 
 @pytest.fixture
 def contracts(
-    deployer, token, forward_state, request_manager_params, fees_params, chain_params, token_params
+    deployer,
+    token,
+    forward_state,
+    request_manager_params,
+    fees_params,
+    chain_params,
+    token_params,
+    local_account,
 ):
     # L1 contracts
     l1_messenger = deployer.deploy(ape.project.TestL1Messenger)
@@ -162,8 +159,8 @@ def contracts(
     request_manager = deployer.deploy(ape.project.RequestManager, *request_manager_params)
 
     # Add allowed LPs
-    fill_manager.addAllowedLp(_LOCAL_ACCOUNT)
-    request_manager.addAllowedLp(_LOCAL_ACCOUNT)
+    fill_manager.addAllowedLp(local_account)
+    request_manager.addAllowedLp(local_account)
 
     # Explicitly allow calls between contracts. The chain of trust:
     #
@@ -202,56 +199,6 @@ def token_list(token, allowance):
 
 
 @pytest.fixture
-def config(request_manager, fill_manager, token, token_list):
-    contracts_info = dict(
-        RequestManager=ContractInfo(
-            deployment_block=BlockNumber(1),
-            address=request_manager.address,
-            abi=[abi.dict() for abi in request_manager.contract_type.abi],
-        ),
-        FillManager=ContractInfo(
-            deployment_block=BlockNumber(1),
-            address=fill_manager.address,
-            abi=[abi.dict() for abi in fill_manager.contract_type.abi],
-        ),
-    )
-    deployment_info = cast(DeploymentInfo, {ape.chain.chain_id: contracts_info})
-    account = eth_account.Account.from_key(_LOCAL_ACCOUNT.private_key)
-    token.mint(account.address, 300)
-    url = URL(ape.config.provider.uri)
-    chains = {}
-    for chain_name in ("l2a", "l2b"):
-        chains[chain_name] = ChainConfig(
-            rpc_url=url,
-            min_source_balance=0,
-            confirmation_blocks=0,
-            poll_period=0.5,
-        )
-
-    config = Config(
-        deployment_info=deployment_info,
-        token_checker=TokenChecker(token_list),
-        account=account,
-        fill_wait_time=0,
-        unsafe_fill_time=600,
-        prometheus_metrics_port=None,
-        base_chain_rpc_url=url,
-        log_level="debug",
-        chains=chains,
-    )
-    beamer.agent.metrics.init(config=config, source_rpc_url=url, target_rpc_url=url)
-    return config
-
-
-@pytest.fixture
-def agent(config):  # pylint:disable=unused-argument
-    agent = Agent(config)
-    agent.start()
-    yield agent
-    agent.stop()
-
-
-@pytest.fixture
 def token(deployer):
     return deployer.deploy(ape.project.MintableToken, int(1e18))
 
@@ -274,18 +221,3 @@ def resolver(contracts):
 @pytest.fixture
 def fill_manager(contracts):
     return contracts.fill_manager
-
-
-@pytest.fixture
-def direction():
-    return TransferDirection(ChainId(ape.chain.chain_id), ChainId(ape.chain.chain_id))
-
-
-@pytest.fixture
-def setup_relayer_executable():
-    relayer = get_relayer_executable()
-    relayer.parent.mkdir(parents=True, exist_ok=True)
-    relayer.write_text("#!/bin/sh\nsleep 5\necho 'hello'")
-    relayer.chmod(relayer.stat().st_mode | stat.S_IEXEC)
-    yield
-    relayer.unlink()
